add support for fuzz testing
CLKNETSIM_FUZZ_MODE variable can be used to enable a fuzz testing mode.
The client doesn't connect to a server and there is no clock control or
networking. When the simulated time reaches CLKNETSIM_FUZZ_START, a
packet read from stdin is forwarded to the CLKNETSIM_FUZZ_PORT port of
the client, and the client is terminated. Packets sent by the client
from the port are written to stdout.