add refclocks based on node clocks
Allow the base of the refclock, against which is specified offset of the
refclock, to be changed to a node. The default is the original base of
the (true) network time.

This enables a simulation of a PHC synchronized by ptp4l and used as a
source by phc2sys in the automatic mode.