bash: remove linker error from logs with valgrind
When using 64-bit valgrind on 32-bit client binaries, new version
of the glibc dynamic linker writes an error to stderr about failed
LD_PRELOAD, which can break tests due to unexpected strings in the log.

Remove the linker error from the logs if valgrind is enabled to avoid
test failures.