add workaround for sanitizer intercepting timer_* functions
The llvm address sanitizer seems to be intercepting timer_settime() and
timer_delete(), which breaks the recently added RTC support using an fd
timer emulating the RTC interrupt.

Provide internal timer functions and use them for fd timers. A different
approach will be needed to fully fix the timer support with the
sanitizer.